refactor: extract desktop cloud promo controller
